t3e

+++ code/modules/mob/living/carbon/superior_animal/robots/robot.dm
@@ -32,8 +32,11 @@
 	viewRange = 8
 	reagent_immune = TRUE
 	toxin_immune = TRUE
+	cold_protection = 1
+	heat_protection = 1
 	var/cleaning = TRUE
 	var/emp_damage = TRUE // Does EMP & Ion weapons cause damage?
+	var/termiation = TRUE
 
 	can_burrow = FALSE
 	colony_friend = FALSE
@@ -104,7 +107,8 @@
 	if(cell_drop)
 		new cell_drop (src.loc)
 		cell_drop = null
-	qdel(src)
+	if(termiation)
+		qdel(src)
 	return
 
 /mob/living/carbon/superior_animal/robot/emp_act(severity)
